requirements.txt
groq
requests
python-dotenv
streamlit